+++ src/EvolutionBack/EvolutionBack/Core/IQueries.cs
@@ -1,5 +1,8 @@
 ﻿namespace EvolutionBack;
 
+/// <summary>
+/// Интерфейс, по которому классы добавляются в DI. Результаты запросов этих классов не должны отслеживаться EntityFrameworkCore
+/// </summary>
 public interface IQueries
 {
 }
