+++ src/EvolutionBack/EvolutionBack/Core/IService.cs
@@ -1,5 +1,8 @@
 ﻿namespace EvolutionBack;
 
+/// <summary>
+/// Интерфейс, по которому классы добавляются в DI.
+/// </summary>
 public interface IService
 {
 }
